move shared constants and update imports
