+++ src/applications/appeals/10182/pages/contestableIssues.js
@@ -4,7 +4,7 @@ import { ContestableIssuesAdditionalInfo } from '../content/contestableIssues';
 
 import { selectionRequired, maxIssues } from '../validations/issues';
 import { hasSomeSelected } from '../utils/helpers';
-import { SELECTED } from '../constants';
+import { SELECTED } from '../../shared/constants';
 
 /**
  * contestable issues with add issue link (list loop)
