+++ src/applications/appeals/10182/tests/nod.cypress.helpers.js
@@ -1,5 +1,5 @@
 import { getDate } from '../utils/dates';
-import { SELECTED } from '../constants';
+import { SELECTED } from '../../shared/constants';
 
 export const getRandomDate = () =>
   getDate({
