+++ src/applications/appeals/10182/tests/pages/issueSummary.unit.spec.jsx
@@ -6,7 +6,8 @@ import sinon from 'sinon';
 import { DefinitionTester } from 'platform/testing/unit/schemaform-utils';
 
 import formConfig from '../../config/form';
-import { SELECTED, CONTESTABLE_ISSUES_PATH } from '../../constants';
+import { CONTESTABLE_ISSUES_PATH } from '../../constants';
+import { SELECTED } from '../../../shared/constants';
 
 describe('NOD selected issues summary page', () => {
   const {
