+++ src/applications/appeals/996/constants.js
@@ -26,9 +26,6 @@ export const BENEFIT_OFFICES_URL = `${HLR_INFO_URL}#find-addresses-for-other-ben
 export const CONTESTABLE_ISSUES_API =
   '/higher_level_reviews/contestable_issues/';
 
-// key for contestedIssues to indicate that the user selected the issue
-export const SELECTED = 'view:selected';
-
 // Including a default until we determine how to get around the user restarting
 // the application after using the "Finish this application later" link
 // See https://dsva.slack.com/archives/C0113MPTGH5/p1600725048027200
@@ -96,38 +93,6 @@ export const CONFERENCE_TIMES_V2 = {
     label: 'Noon to 4:30 p.m. ET',
     submit: '1200-1630 ET',
   },
-};
-
-// Values from Lighthouse maintained schema v2
-// see https://github.com/department-of-veterans-affairs/vets-api/blob/master/modules/appeals_api/config/schemas/v2/200996.json
-export const MAX_LENGTH = {
-  SELECTIONS: 100, // submitted issues (not in schema)
-  ISSUE_NAME: 140,
-  DISAGREEMENT_REASON: 90,
-  EMAIL: 255,
-  COUNTRY_CODE: 3,
-  AREA_CODE: 4,
-  PHONE_NUMBER: 14,
-  PHONE_NUMBER_EXT: 10,
-  ADDRESS_LINE1: 60,
-  ADDRESS_LINE2: 30,
-  ADDRESS_LINE3: 10,
-  CITY: 60,
-  COUNTRY: 2,
-  ZIP_CODE5: 5,
-  POSTAL_CODE: 16,
-  REP_FIRST_NAME: 30,
-  REP_LAST_NAME: 40,
-};
-
-// Using MAX_LENGTH.DISAGREEMENT_REASON (90) and with all checkboxes selected,
-// this string is submitted - the numbers constitute the "something else" typed
-// in value
-// "service connection,effective date,disability evaluation,1234567890123456789012345678901234"
-export const SUBMITTED_DISAGREEMENTS = {
-  serviceConnection: 'service connection',
-  effectiveDate: 'effective date',
-  evaluation: 'disability evaluation',
 };
 
 export const CONTESTABLE_ISSUES_PATH = 'contestable-issues'; // v2
