+++ src/applications/appeals/996/pages/areaOfDisagreement.js
@@ -10,7 +10,7 @@ import {
 import { areaOfDisagreementRequired } from '../validations/issues';
 import { calculateOtherMaxLength } from '../utils/disagreement';
 import { getIssueName } from '../utils/helpers';
-import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../constants';
+import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../../shared/constants';
 
 // add 1 for last comma
 const allDisagreementsLength =
