+++ src/applications/appeals/996/tests/containers/ConfirmationPage.unit.spec.jsx
@@ -7,7 +7,8 @@ import formConfig from '../../config/form';
 import initialData from '../schema/initialData';
 
 import ConfirmationPage from '../../containers/ConfirmationPage';
-import { WIZARD_STATUS, SELECTED, SAVED_CLAIM_TYPE } from '../../constants';
+import { WIZARD_STATUS, SAVED_CLAIM_TYPE } from '../../constants';
+import { SELECTED } from '../../../shared/constants';
 
 const data = {
   user: {
