+++ src/applications/appeals/996/utils/disagreement.js
@@ -1,4 +1,4 @@
-import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../constants';
+import { MAX_LENGTH, SUBMITTED_DISAGREEMENTS } from '../../shared/constants';
 import { getIssueName } from './helpers';
 
 /**
